\name{EnquireR-internal}

\alias{barplot_function}
\alias{barplot_sweave}
\alias{barplotclustervsvar}
\alias{barplotgroupes}
\alias{chisq.function}
\alias{chisq.sweave}
\alias{col.barplot}
\alias{cuttree.graph}
\alias{legend2}
\alias{MCA2}
\alias{tab_chi2}
\alias{catdes2}

\title{Internal EnquireR objects}
\description{Internal EnquireR objects.}
\details{These are not to be called by the user.}
\keyword{internal}