\name{EnquireR-package}
\alias{EnquireR-package}
\alias{EnquireR}
\docType{package}
\title{
A package dedicated to questionnaires
}
\description{
The EnQuireR package focuses on categorical variables and provides many tools to
automate the survey process. It includes both univariate and multivariate data analyses
comprising Multiple Correspondence Analysis (MCA), clustering analysis and semantic
marking. The package also offers an easier view of the results by the automatic generation
of a .pdf report and of a Beamer type presentation via the use of Sweave. In order to generate those report and presentation, you will need MikTeX and TeXnicCenter. To
facilitate the installation for the user, the .sty and .cls files needed for the generation of the
reports have been integrated to the package; hence its size.
}
\details{
\tabular{ll}{
Package: \tab EnQuireR\cr
Type: \tab Package\cr
Version: \tab 0.09\cr
Date: \tab 2009-06-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sebastien Le \email{sebastien.le@agrocampus-ouest.fr} \cr
Jerome Bouche \email{jerome.bouche@yahoo.fr} \cr
Marine Cadoret \email{marine.cadoret@agrocampus-ouest.fr} \cr
Gwenaelle Fournier \email{gwenaelle.fournier@gmail.com} \cr
Olivier Fournier \email{olivier.fournier76@gmail.com} \cr
Francois Le Poder \email{francois.lepoder@gmail.com}
}


\keyword{ package }


