\name{tea}
\alias{tea}
\docType{data}
\title{tea (data)}
\description{
The data used here concern a questionnaire on tea. We asked to 300 individuals how they drink tea (19 questions), what are their product's perception (12 questions) and some personal details (4 questions). 
}
\usage{data(tea)}
\format{
A data frame with 300 rows and 35 columns. Rows represent the individuals, columns represent the different questions.
}


\examples{
\dontrun{
data(tea)
}
}

